feat(chord chart): title and position of a chord can be provided as part of the chord
Before this change both title and positions were part of the svguitar configuration object. With
this change, the title and position can be provided in the chord object. If title and position are
provided in both configuration and chart, the position and title of the configuration will be
ignored. So with this change the title and position provided in the configuration will act as
defaults for all chords created with the configured instance.

fix #43