fix(chord settings): rename nutStrokeColor and nutStrokeWidth to strokeColor and strokeWidth
To be consistent with other options that are set for individual notes, the settings to change the
stroke have been renamed to strokeWidth and strokeColor. This is a breaking change but since the
previous version hasn't been released for more than 24 hours I'm making this change without a major
version bump.