refactor: rename nut to finger everywhere
Due to a lack of guitar knowledge of the maintainer of this library the fingers were wrongly called
"nuts" across the entire codebase, including the API. Since a feature is planned to configure the
actual nut (with properties such as nutSize and nutColor etc) the decision was made to introduce a
breaking change and change all wrongly named nut* properties to finger*.

BREAKING CHANGE: rename nut to finger everywhere. Specifically the configuration properties nutSize,
nutColor, nutTextColor, nutTextSize, nutStrokeColor, nutStrokeWidth are now renamed to fingerSize,
fingerColor, fingerTextColor, fingerTextSize, fingerStrokeColor, fingerStrokeWidth.